Consolidate fragmented bitcoin trading scripts and update imports in tests
