+++ tests/test_bitcoin_simulation_actions.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pytest
-from bitcoin_simulation import _generate_actions
+from bitcoin_trading import _generate_actions
 
 def test_generate_actions_empty():
     prices = np.array([])
